Add programId to CreateGiftCardRequest If a AGCOD_PROGRAMID env variable exists #12
The program ID is an approved identifier provided by Amazon through a submission process. Partners will submit client and use case information through the portal in the Payment Tracker. Approved submissions will receive a reference number that will be added to each transaction call to the API. The programID is alphanumeric and can be up to 100 characters in length. The programId is a required field.